requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
Pillow==7.1.1
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
